+++ pytest_django/fixtures.py
@@ -3,10 +3,10 @@
 from __future__ import annotations
 
 import os
-from collections.abc import Generator, Iterable, Sequence
+from collections.abc import Callable, Generator, Iterable, Sequence
 from contextlib import AbstractContextManager, contextmanager
 from functools import partial
-from typing import TYPE_CHECKING, Protocol
+from typing import TYPE_CHECKING, Any, Protocol
 
 import pytest
 
@@ -16,7 +16,7 @@ from .lazy_django import skip_if_no_django
 
 
 if TYPE_CHECKING:
-    from typing import Any, Callable, Literal, Optional, Union
+    from typing import Literal, TypeAlias
 
     import django
     import django.test
@@ -24,10 +24,10 @@ if TYPE_CHECKING:
     from . import DjangoDbBlocker
     from .django_compat import _User, _UserModel
 
-    _DjangoDbDatabases = Optional[Union[Literal["__all__"], Iterable[str]]]
-    _DjangoDbAvailableApps = Optional[list[str]]
+    _DjangoDbDatabases: TypeAlias = Literal["__all__"] | Iterable[str] | None
+    _DjangoDbAvailableApps: TypeAlias = list[str] | None
     # transaction, reset_sequences, databases, serialized_rollback, available_apps
-    _DjangoDb = tuple[bool, bool, _DjangoDbDatabases, bool, _DjangoDbAvailableApps]
+    _DjangoDb: TypeAlias = tuple[bool, bool, _DjangoDbDatabases, bool, _DjangoDbAvailableApps]
 
 
 __all__ = [
@@ -124,11 +124,11 @@ def _get_databases_for_test(test: pytest.Item) -> tuple[Iterable[str], bool]:
         marker_db = test.get_closest_marker("django_db")
         if marker_db:
             (
-                transaction,
-                reset_sequences,
+                _transaction,
+                _reset_sequences,
                 databases,
                 serialized_rollback,
-                available_apps,
+                _available_apps,
             ) = validate_django_db(marker_db)
         elif "db" in fixtures or "transactional_db" in fixtures or "live_server" in fixtures:
             serialized_rollback = "django_db_serialized_rollback" in fixtures
@@ -160,7 +160,7 @@ def _get_databases_for_setup(
 
 
 @pytest.fixture(scope="session")
-def django_db_setup(
+def django_db_setup(  # noqa: PLR0917
     request: pytest.FixtureRequest,
     django_test_environment: None,  # noqa: ARG001
     django_db_blocker: DjangoDbBlocker,
@@ -168,7 +168,7 @@ def django_db_setup(
     django_db_keepdb: bool,
     django_db_createdb: bool,
     django_db_modify_db_settings: None,  # noqa: ARG001
-) -> Generator[None, None, None]:
+) -> Generator[None]:
     """Top level fixture to ensure test databases are available"""
     from django.test.utils import setup_databases, teardown_databases
 
@@ -208,10 +208,12 @@ def _django_db_helper(
     request: pytest.FixtureRequest,
     django_db_setup: None,  # noqa: ARG001
     django_db_blocker: DjangoDbBlocker,
-) -> Generator[None, None, None]:
+) -> Generator[None]:
     if is_django_unittest(request):
         yield
         return
+
+    from django import VERSION
 
     marker = request.node.get_closest_marker("django_db")
     if marker:
@@ -289,7 +291,14 @@ def _django_db_helper(
         PytestDjangoTestCase.setUpClass()
 
         test_case = PytestDjangoTestCase(methodName="__init__")
-        test_case._pre_setup()
+        if VERSION >= (6, 0):
+            pre_setup_ran_eagerly = PytestDjangoTestCase._pre_setup_ran_eagerly
+        else:
+            pre_setup_ran_eagerly = getattr(PytestDjangoTestCase, "_pre_setup_ran_eagerly", False)
+        if not pre_setup_ran_eagerly:
+            # For a TransactionTestCase, setUpClass() has already run _pre_setup() and set
+            # this flag to say so.
+            test_case._pre_setup()
 
         yield
 
@@ -536,8 +545,20 @@ def async_rf() -> django.test.AsyncRequestFactory:
     return AsyncRequestFactory()
 
 
-class SettingsWrapper:
-    def __init__(self) -> None:
+class Settings:
+    """The type of the :fixture:`settings` fixture.
+
+    Allows getting, setting and deleting Django settings for the duration of the test.
+    """
+
+    def __init__(
+        self,
+        *,
+        _is_pytest_django: bool = False,
+    ) -> None:
+        assert _is_pytest_django, (
+            "Settings should only be instantiated from the `settings` fixture"
+        )
         self._to_restore: list[django.test.override_settings]
         object.__setattr__(self, "_to_restore", [])
 
@@ -564,7 +585,7 @@ class SettingsWrapper:
 
         return getattr(settings, attr)
 
-    def finalize(self) -> None:
+    def _finalize(self) -> None:
         for override in reversed(self._to_restore):
             override.disable()
 
@@ -572,19 +593,19 @@ class SettingsWrapper:
 
 
 @pytest.fixture
-def settings() -> Generator[SettingsWrapper, None, None]:
+def settings() -> Generator[Settings]:
     """A Django settings object which restores changes after the testrun"""
     skip_if_no_django()
 
-    wrapper = SettingsWrapper()
+    wrapper = Settings(_is_pytest_django=True)
     yield wrapper
-    wrapper.finalize()
+    wrapper._finalize()
 
 
 @pytest.fixture(scope="session")
 def live_server(
     request: pytest.FixtureRequest,
-) -> Generator[live_server_helper.LiveServer, None, None]:
+) -> Generator[live_server_helper.LiveServer]:
     """Run a live Django server in the background during tests
 
     The address the server is started from is taken from the
@@ -617,7 +638,7 @@ def live_server(
 
 
 @pytest.fixture(autouse=True)
-def _live_server_helper(request: pytest.FixtureRequest) -> Generator[None, None, None]:
+def _live_server_helper(request: pytest.FixtureRequest) -> Generator[None]:
     """Helper to make live_server work, internal to pytest-django.
 
     This helper will dynamically request the transactional_db fixture
@@ -668,7 +689,7 @@ def _assert_num_queries(
     info: str | None = None,
     *,
     using: str | None = None,
-) -> Generator[django.test.utils.CaptureQueriesContext, None, None]:
+) -> Generator[django.test.utils.CaptureQueriesContext]:
     from django.db import connection as default_conn, connections
     from django.test.utils import CaptureQueriesContext