+++ tests/test_django_configurations.py
@@ -44,8 +44,10 @@ def test_dc_env(pytester: pytest.Pytester, monkeypatch: pytest.MonkeyPatch) -> N
     result = pytester.runpytest_subprocess()
     result.stdout.fnmatch_lines(
         [
-            "django: version: *, settings: tpkg.settings_env (from env), "
-            "configuration: MySettings (from env)",
+            (
+                "django: version: *, settings: tpkg.settings_env (from env), "
+                "configuration: MySettings (from env)"
+            ),
             "* 1 passed*",
         ]
     )
@@ -78,8 +80,10 @@ def test_dc_env_overrides_ini(pytester: pytest.Pytester, monkeypatch: pytest.Mon
     result = pytester.runpytest_subprocess()
     result.stdout.fnmatch_lines(
         [
-            "django: version: *, settings: tpkg.settings_env (from env), "
-            "configuration: MySettings (from env)",
+            (
+                "django: version: *, settings: tpkg.settings_env (from env), "
+                "configuration: MySettings (from env)"
+            ),
             "* 1 passed*",
         ]
     )
@@ -111,8 +115,10 @@ def test_dc_ini(pytester: pytest.Pytester, monkeypatch: pytest.MonkeyPatch) -> N
     result = pytester.runpytest_subprocess()
     result.stdout.fnmatch_lines(
         [
-            "django: version: *, settings: tpkg.settings_ini (from ini), "
-            "configuration: MySettings (from ini)",
+            (
+                "django: version: *, settings: tpkg.settings_ini (from ini), "
+                "configuration: MySettings (from ini)"
+            ),
             "* 1 passed*",
         ]
     )
@@ -145,8 +151,10 @@ def test_dc_option(pytester: pytest.Pytester, monkeypatch: pytest.MonkeyPatch) -
     result = pytester.runpytest_subprocess("--ds=tpkg.settings_opt", "--dc=MySettings")
     result.stdout.fnmatch_lines(
         [
-            "django: version: *, settings: tpkg.settings_opt (from option),"
-            " configuration: MySettings (from option)",
+            (
+                "django: version: *, settings: tpkg.settings_opt (from option),"
+                " configuration: MySettings (from option)"
+            ),
             "* 1 passed*",
         ]
     )