+++ tests/test_fixtures.py
@@ -24,18 +24,22 @@ from django.utils.encoding import force_str
 
 from .helpers import DjangoPytester
 
-from pytest_django import DjangoAssertNumQueries, DjangoCaptureOnCommitCallbacks, DjangoDbBlocker
+from pytest_django import (
+    DjangoAssertNumQueries,
+    DjangoCaptureOnCommitCallbacks,
+    DjangoDbBlocker,
+    Settings,
+)
 from pytest_django_test.app.models import Item
 
 
 if TYPE_CHECKING:
     from pytest_django.django_compat import _User, _UserModel
-    from pytest_django.fixtures import SettingsWrapper
     from pytest_django.live_server_helper import LiveServer
 
 
 @contextmanager
-def nonverbose_config(config: pytest.Config) -> Generator[None, None, None]:
+def nonverbose_config(config: pytest.Config) -> Generator[None]:
     """Ensure that pytest's config.option.verbose is <= 0."""
     if config.option.verbose <= 0:
         yield
@@ -139,8 +143,10 @@ def test_django_assert_max_num_queries_db(
                 Item.objects.create(name="3-quux")
 
         assert excinfo.value.args == (
-            "Expected to perform 2 queries or less but 3 were done "
-            "(add -v option to show queries)",
+            (
+                "Expected to perform 2 queries or less but 3 were done "
+                "(add -v option to show queries)"
+            ),
         )
         assert len(captured.captured_queries) == 3
         assert "1-foo" in captured.captured_queries[0]["sql"]
@@ -342,40 +348,40 @@ def test_django_capture_on_commit_callbacks_transactional(
 class TestSettings:
     """Tests for the settings fixture, order matters"""
 
-    def test_modify_existing(self, settings) -> None:
+    def test_modify_existing(self, settings: Settings) -> None:
         assert settings.SECRET_KEY == "foobar"
         assert real_settings.SECRET_KEY == "foobar"
         settings.SECRET_KEY = "spam"
         assert settings.SECRET_KEY == "spam"
         assert real_settings.SECRET_KEY == "spam"
 
-    def test_modify_existing_again(self, settings) -> None:
+    def test_modify_existing_again(self, settings: Settings) -> None:
         assert settings.SECRET_KEY == "foobar"
         assert real_settings.SECRET_KEY == "foobar"
 
-    def test_new(self, settings) -> None:
+    def test_new(self, settings: Settings) -> None:
         assert not hasattr(settings, "SPAM")
         assert not hasattr(real_settings, "SPAM")
         settings.SPAM = "ham"
         assert settings.SPAM == "ham"
         assert real_settings.SPAM == "ham"
 
-    def test_new_again(self, settings) -> None:
+    def test_new_again(self, settings: Settings) -> None:
         assert not hasattr(settings, "SPAM")
         assert not hasattr(real_settings, "SPAM")
 
-    def test_deleted(self, settings) -> None:
+    def test_deleted(self, settings: Settings) -> None:
         assert hasattr(settings, "SECRET_KEY")
         assert hasattr(real_settings, "SECRET_KEY")
         del settings.SECRET_KEY
         assert not hasattr(settings, "SECRET_KEY")
         assert not hasattr(real_settings, "SECRET_KEY")
 
-    def test_deleted_again(self, settings) -> None:
+    def test_deleted_again(self, settings: Settings) -> None:
         assert hasattr(settings, "SECRET_KEY")
         assert hasattr(real_settings, "SECRET_KEY")
 
-    def test_signals(self, settings) -> None:
+    def test_signals(self, settings: Settings) -> None:
         result = []
 
         def assert_signal(
@@ -474,13 +480,6 @@ class TestLiveServer:
         TestLiveServer._test_settings_before_run = True  # type: ignore[attr-defined]
 
     def test_url(self, live_server: LiveServer) -> None:
-        assert live_server.url == force_str(live_server)
-
-    def test_change_settings(
-        self,
-        live_server: LiveServer,
-        settings: SettingsWrapper,  # noqa: ARG002
-    ) -> None:
         assert live_server.url == force_str(live_server)
 
     @pytest.mark.skipif("PYTEST_XDIST_WORKER" in os.environ, reason="xdist in use")
@@ -766,7 +765,7 @@ class Test_django_db_blocker:
     def test_block_manually(self, django_db_blocker: DjangoDbBlocker) -> None:
         try:
             django_db_blocker.block()
-            with pytest.raises(RuntimeError, match="^Database access not allowed,"):
+            with pytest.raises(RuntimeError, match=r"^Database access not allowed,"):
                 Item.objects.exists()
         finally:
             django_db_blocker.restore()
@@ -774,7 +773,7 @@ class Test_django_db_blocker:
     @pytest.mark.django_db
     def test_block_with_block(self, django_db_blocker: DjangoDbBlocker) -> None:
         with django_db_blocker.block():
-            with pytest.raises(RuntimeError, match="^Database access not allowed,"):
+            with pytest.raises(RuntimeError, match=r"^Database access not allowed,"):
                 Item.objects.exists()
 
     def test_unblock_manually(self, django_db_blocker: DjangoDbBlocker) -> None: